requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.3
sqlparse==0.4.2
XlsxWriter==3.0.2
